backend/requirements.txt
Django~=5.2.4
djangorestframework~=3.16.0
python-dotenv~=1.1.1